requirements.txt
Django >= 5.0.3, < 5.1.0
djangorestframework >= 3.14.0, < 3.15.0
psycopg2>=2.9,<2.9.9
drf-spectacular>=0.27.0,<0.27.1
Pillow>=8.3.0,<8.4.0